añadido algolia, y reestructurada  carpeta de functionss. Arreglado seguir lugar
