+++ functions/.env.listopic
@@ -1,1 +1,4 @@
 GOOGLE_PLACES_API_KEY="AIzaSyBV26T665w65z1FNdDk8mwHSH1o0c2PeW8"
+
+ALGOLIA_APP_ID="FI4Q0XQABV"
+ALGOLIA_API_KEY="ed3bb83496e3dd40743624a90991c51c"
