+++ functions/package.json
@@ -14,9 +14,10 @@
   },
   "main": "index.js",
   "dependencies": {
+    "algoliasearch": "^4.24.0",
     "cors": "^2.8.5",
-    "firebase-admin": "^12.0.0",
-    "firebase-functions": "^5.0.0",
+    "firebase-admin": "^13.4.0",
+    "firebase-functions": "^6.4.0",
     "node-fetch": "^2.7.0"
   },
   "devDependencies": {
